requirements.txt
requests
beautifulsoup4
faiss-cpu
sentence-transformers
numpy
fpdf2
tk
torch